Fixed sucessors function
During the first iteration it was setting start and end locations for
the new updated state wrong.  When picking up packages - it is supposed
to set the current vehicle location to the same location of the package
we are picking up but we were setting it to the location of the next
package we were supposed to pick up.  Same mistake was made in the
dropping off packages - we were setting the current location to be the
location of the next drop location instead of where we just dropped a
package to